% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aucPlot}
\alias{aucPlot}
\title{Sample, Train, and Predict logistic regression
model using singleCellExperiment and glmnet}
\usage{
aucPlot(casc_list)
}
\arguments{
\item{casc_list}{A casc_list object produced by casc}
}
\value{
A ggplot scatterplot with AUCs plotted for each cluster
}
\description{
Sample, Train, and Predict logistic regression
model using singleCellExperiment and glmnet
}
\examples{
library(SingleCellExperiment)

counts <- matrix(rnorm(40000, 10, 10), ncol=200, nrow=200)
sce <- SingleCellExperiment(assays = list(logcounts = counts))
colnames(sce) <- stringi::stri_rand_strings(200, 5)
rownames(sce) <- stringi::stri_rand_strings(200, 5)

cluster_1 <- rep(c(0, 1, 1, 1, 1, 0, 1, 1, 1, 1), 20)
cluster_2 <- rep(c(0, 1, 1, 0, 1, 0, 1, 1, 0, 1), 20)

casc_list <- casc(sce, list(cluster_1, cluster_2), marker_num=1500)
aucPlot(casc_list)

}
