% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{avgAUC}
\alias{avgAUC}
\title{Finds the average AUC from a list of roc objects}
\usage{
avgAUC(roc_l)
}
\arguments{
\item{roc_l}{A list of casc objects produced by casc}
}
\value{
Average auc for the list, numeric
}
\description{
Finds the average AUC from a list of roc objects
}
