% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casc_alpha.R
\name{casc_alpha}
\alias{casc_alpha}
\title{Covariate-assisted spectural clustering using fixed weight alpha}
\usage{
casc_alpha(A, X, L.tau, K, alpha, method, randStarts)
}
\arguments{
\item{A}{an adjacent matrix of the nodes, n-by-n}

\item{X}{covariate matrix, n-by-p}

\item{alpha}{the weight of the XX' matrix to be added to regularized
Laplacian. Default value 100}

\item{method}{xlustering method. See details}
}
\value{
The clustering assignments along with the within-cluster sum of
  squares.
}
\description{
\code{casc_alpha} returns the covariate-assisted pectrual clustering
assignments with fixed alpha weight.
}
\examples{

\dontrun{
sum("a")
}
}

