% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cascer}
\alias{cascer}
\title{Sample, Train, and Predict logistic regression model using singleCellExperiment and glmnet}
\usage{
cascer(sce, clusters)
}
\arguments{
\item{sce}{a singleCellExperiment}

\item{clusters}{A cluster, an array or list of integers of same length as number of cells in sce}
}
\value{
A casc object with predicted classes, auc, response, and truths
}
\description{
Sample, Train, and Predict logistic regression model using singleCellExperiment and glmnet
}
