% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASpecClust.R
\name{CASpecClust}
\alias{CASpecClust}
\title{Covariate-assisted Spectral Clustering}
\usage{
CASpecClust(A, X = NULL, K = 2, n.alpha = 100, center = FALSE,
  method = 3, randStarts = 20)
}
\arguments{
\item{A}{an adjacency matrix of the nodes, n by n}

\item{X}{covariate matrix, n by p.
Categorical variables should be re-expressed with dummy variables using
one-hot encoding, see vignettes for example.}

\item{K}{number of clusters}

\item{n.alpha}{number of alpha values in the valid range to search. Defaults
to 100}

\item{center}{whether to center the covariates X. Defaults to FALSE}

\item{method}{clustering method. See details}

\item{randStarts}{number of random starts in K-means. Defaults to 20}
}
\value{
A list of three elememts:
\enumerate{
  \item alpha: values of tuning parameter in the valid range
  \item WCSS: within-cluster sum of squares corresponding to the alpha values
  \item cluster: the optimal clustering assignments
}
}
\description{
\code{CASpecClust} returns the covariate-assisted spectral clustering assignments.
This is a function that implements the methods described in Binkiewicz et.al. (2016).
\code{CASpecClust} takes the adjacency matrix \code{A}, covariate matrix \code{X},
calculates a similarity matrix and peforms spectral clustering.
}
\details{
The adjacency matrix and covariate matrices are first converted into sparse format
to calculate the graph laplacian:
 \deqn{L_\tau = D^{−1/2}_\tau A D^{−1/2}_\tau}

spectral clusterging is performed on a n-by-n "similarity matrix", which is defined
according to method used:

\enumerate{
  \item regularized spectral clustering:
      \deqn{L_\tau}
  \item assortative covariate-assisted spectral clustering:
      \deqn{L_\tau + \alpha XX^t}
  \item covariate-assisted spectral clustering:
      \deqn{L_\tau L_\tau + \alpha XX^t}
}

The nodes are clustered into \code{K} clusters.
}
\examples{
See vignettes

}

