% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTuningRange.R
\name{getTuningRange}
\alias{getTuningRange}
\title{Get range of tuning weight alpha}
\usage{
getTuningRange(L.tau, X, K, method)
}
\arguments{
\item{X}{covariate matrix, n-by-p}

\item{K}{number of clusters}

\item{method}{}

\item{A}{an adjacent matrix of the nodes, n-by-n}
}
\value{
the max and min of possible alpha values
}
\description{
\code{getTuningRange} determins the valid range of tuning variable alpha to
search over
}
\examples{

\dontrun{
sum("a")
}
}

