% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multROC}
\alias{multROC}
\title{Sample, Train, and Predict logistic regression model using singleCellExperiment and glmnet}
\usage{
multROC(casc)
}
\arguments{
\item{casc}{A casc object produced by cascer}
}
\value{
A ggplot object with ROC curves plotted for each cluster
}
\description{
Sample, Train, and Predict logistic regression model using singleCellExperiment and glmnet
}
