% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multROCPlot}
\alias{multROCPlot}
\title{Plot multiple roc objects on the same ggplot}
\usage{
multROCPlot(casc)
}
\arguments{
\item{casc}{A casc object}
}
\value{
A ggplot object with ROC curves plotted for each cluster
}
\description{
Plot multiple roc objects on the same ggplot
}
\examples{
library(SingleCellExperiment)

counts <- matrix(rnorm(40000, 10, 10), ncol=200, nrow=200)
sce <- SingleCellExperiment(assays = list(logcounts = counts))
colnames(sce) <- stringi::stri_rand_strings(200, 5)
rownames(sce) <- stringi::stri_rand_strings(200, 5)

cluster_1 <- rep(c(0, 1, 1, 1, 1, 0, 1, 1, 1, 1), 20)
cluster_2 <- rep(c(0, 1, 1, 0, 1, 0, 1, 1, 0, 1), 20)

casc_list <- casc(sce, list(cluster_1, cluster_2), marker_num=1500)
multROCPlot(casc_list$casc_1)

}
