% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleSplit}
\alias{sampleSplit}
\title{Splits SingleCellExperiment object into test and training data}
\usage{
sampleSplit(sce, marker_num = 2000)
}
\value{
a list, including both the training and test splits
}
\description{
Splits SingleCellExperiment object into test and training data
}
\keyword{internal}
