% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{sce_sim}
\alias{sce_sim}
\title{a simulated SingleCellExpermient object for vignette with 1000 cells
and 500 genes. 10 artificial clusters contain 20 marker genes a piece.}
\format{SingleCellExperiment}
\usage{
data(sce_sim)
}
\description{
a simulated SingleCellExpermient object for vignette with 1000 cells
and 500 genes. 10 artificial clusters contain 20 marker genes a piece.
}
\keyword{datasets}
