% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{single_casc}
\alias{single_casc}
\title{Sample, Train, and Predict logistic regression model
using singleCellExperiment and glmnet, one object at a time}
\usage{
single_casc(cluster_name, dataSplits, alpha = 0.5)
}
\value{
a casc object
}
\description{
Sample, Train, and Predict logistic regression model
using singleCellExperiment and glmnet, one object at a time
}
\keyword{internal}
