% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{test_casc_list}
\alias{test_casc_list}
\title{casc objects within a casc_list object for testing purposes}
\format{A casc_list obj with two casc objects within}
\usage{
data(test_casc_list)
}
\description{
casc objects within a casc_list object for testing purposes
}
\keyword{datasets}
