% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{test_sce}
\alias{test_sce}
\title{a SingleCellExpermient object for testing purposes}
\format{SingleCellExperiment}
\usage{
data(test_sce)
}
\description{
a SingleCellExpermient object for testing purposes
}
\keyword{datasets}
