% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variableGenes}
\alias{variableGenes}
\title{Calculate most variable genes for SingleCellExperiment object}
\usage{
variableGenes(sce, marker_num)
}
\value{
sce with top marker_num most variable genes
}
\description{
Calculate most variable genes for SingleCellExperiment object
}
\keyword{internal}
