% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multROC}
\alias{multROC}
\title{calculate roc objects from a casc object, one for each cluster}
\usage{
multROC(truths, response)
}
\arguments{
\item{truths}{List of original class labels}

\item{response}{matrix or data.frame of probabilities of each class}
}
\value{
A list of roc objects
}
\description{
calculate roc objects from a casc object, one for each cluster
}
