% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.casc}
\alias{print.casc}
\title{Print a \code{casc}}
\usage{
\method{print}{casc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{casc}}

\item{...}{Additional arguments (unused)}
}
\value{
Prints a structured representation of the \code{casc}
}
\description{
Print a \code{casc}
}
